requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# LINE Bot SDK
line-bot-sdk==3.11.0

# Notion Client
notion-client==2.2.1

# Anthropic Claude
anthropic==0.39.0

# Settings Management
pydantic-settings==2.5.2

# HTTP Client
httpx==0.27.2

# Environment
python-dotenv==1.0.1
